Version 210210.1 Uploaded
Changed the TagManager.asset from the project. Cleaned unused layers and left "Collider" layer for debugging purposes in layer index 23.

Why 23?
The previous (now unused) layers are used by the VRChat SDK, as we don't need a specific layer index for debugging colliders at Decentraland it is better to have it after the ones used by VRChat in case we want to migrate any DCL project to the VRChat Ecosystem